uhttpd: use P-256 and 10 years validity for certs
The uhttpd package takes care of creating self-signed certificates if
px5g is installed. This improves the security of router management as it
encrypts the LuCI connection.

Once it's enabled by default the certificates should life longer than
only two years, as most OpenWrt devices run longer and sysupgrades do
not trigger certificate regeneration. For that reason bump the default
valid days to 10 years.

The EC P-256 curve is faster than RSA which which improves the user
experience on embedded devices. EC P-256 is support for as old devices
as Android 4.4.

Signed-off-by: Paul Spooren <mail@aparcar.org>